requirements.txt
argcomplete==3.1.6
blinker==1.7.0
boto3==1.33.1
botocore==1.33.1
certifi==2023.11.17
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
durationpy==0.5
Flask==3.0.0
hjson==3.1.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.3
placebo==0.9.0
pluggy==1.3.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.8.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.5.2
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
zappa==0.58.0
